Refactor GitHub connector: separate webhook functionality into dedicated file
- Extract webhook-related code from github.go to webhook.go
- Move webhook signature validation, event processing, and HTTP server logic
- Maintain all existing functionality and interfaces
- Keep buildWebhookURL in github.go as it's used by ConfigureWebhooks
- Improve code organization and maintainability
